This is a massive fucking hack.
Due to the bullshit in the libretro mupen core, spurious buffer swaps corrupt
dear-imgui. So now to compensate for the fucking shit API, we have to weave UI
code to buffer swap cleanly at the same rate, until a better means of working
with the core is done, or another core is done entirely that behaves with how
the fuck I wanna do this. Spent all morning debugging to find this shit.

Render to textures/image buffer handles FFS instead of wanting to swap willy
nilly!